requirements.txt
flask
schedule
python-dotenv
openai
google-generativeai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
twilio
